Updated CreateFile added metrics restructured packages
